+++ src/formatter/helpers/event_data_collector.js
