+++ src/formatter/helpers/gherkin_document_parser.js
