+++ src/formatter/helpers/index.js
@@ -1,5 +1,11 @@
+import * as GherkinDocumentParser from './gherkin_document_parser'
+import * as PickleParser from './pickle_parser'
+
+export { default as EventDataCollector } from './event_data_collector'
+export { default as KeywordType, getStepKeywordType } from './keyword_type'
 export { formatError } from './error_helpers'
 export { formatIssue, isIssue } from './issue_helpers'
 export { formatLocation } from './location_helpers'
 export { formatSummary } from './summary_helpers'
 export { getUsage } from './usage_helpers'
+export { GherkinDocumentParser, PickleParser }
