+++ src/formatter/helpers/keyword_type.js
