+++ src/formatter/helpers/keyword_type_spec.js
