+++ src/formatter/helpers/pickle_parser.js
