+++ src/formatter/progress_bar_formatter_spec.js
@@ -3,7 +3,7 @@ import ProgressBarFormatter from './progress_bar_formatter'
 import Status from '../status'
 import { EventEmitter } from 'events'
 import Gherkin from 'gherkin'
-import EventDataCollector from '../event_data_collector'
+import { EventDataCollector } from './helpers'
 
 describe('ProgressBarFormatter', function() {
   beforeEach(function() {
