+++ src/index.js
@@ -1,9 +1,6 @@
 // Top level
 export { default as Cli } from './cli'
-export { default as GherkinDocumentParser } from './gherkin_document_parser'
-export { default as KeywordType, getStepKeywordType } from './keyword_type'
 export { default as PickleFilter } from './pickle_filter'
-export { default as PickleParser } from './pickle_parser'
 export { default as Runtime } from './runtime'
 export { default as Status } from './status'
 export {
